requirements.txt
python-telegram-bot[ext]>=20.7
httpx>=0.27.0
python-dotenv>=1.0.0
